SVC: move function_wrappers.h into kernel
It is only used by kernel/SVC
