+++ src/core/CMakeLists.txt
@@ -103,7 +103,6 @@ add_library(core STATIC
     hle/applets/mint.h
     hle/applets/swkbd.cpp
     hle/applets/swkbd.h
-    hle/function_wrappers.h
     hle/ipc.h
     hle/ipc_helpers.h
     hle/kernel/address_arbiter.cpp
@@ -117,6 +116,7 @@ add_library(core STATIC
     hle/kernel/errors.h
     hle/kernel/event.cpp
     hle/kernel/event.h
+    hle/kernel/function_wrappers.h
     hle/kernel/handle_table.cpp
     hle/kernel/handle_table.h
     hle/kernel/hle_ipc.cpp
