+++ src/core/hle/kernel/function_wrappers.h
