+++ src/core/hle/kernel/svc.cpp
@@ -11,12 +11,12 @@
 #include "common/scope_exit.h"
 #include "core/arm/arm_interface.h"
 #include "core/core_timing.h"
-#include "core/hle/function_wrappers.h"
 #include "core/hle/kernel/address_arbiter.h"
 #include "core/hle/kernel/client_port.h"
 #include "core/hle/kernel/client_session.h"
 #include "core/hle/kernel/errors.h"
 #include "core/hle/kernel/event.h"
+#include "core/hle/kernel/function_wrappers.h"
 #include "core/hle/kernel/handle_table.h"
 #include "core/hle/kernel/ipc.h"
 #include "core/hle/kernel/memory.h"
